+++ changelog.md
@@ -4,7 +4,7 @@
 - custom mobile labels bug fixed in default locale.
 - monthYearSeparator added. [#98-1](https://github.com/shahabyazdi/react-multi-date-picker/issues/98#issue-1127212098)
 - formatMonth & formatYear added [#98-2](https://github.com/shahabyazdi/react-multi-date-picker/issues/98#issuecomment-1110705026)
-- now onChange method has (date, validatedValue, input, isTyping) arguments for validations.
+- in datepicker onChange method has (date, {validatedValue, input, isTyping}) arguments for validations.
 - highlightToday added. [#145](https://github.com/shahabyazdi/react-multi-date-picker/issues/145)
 - dateSeparator added.